extract accounts mod into a file
