+++ pallets/order-book/src/test_utils/accounts.rs
@@ -1,0 +1,33 @@
+use codec::Decode;
+
+pub fn alice<T: frame_system::Config>() -> <T as frame_system::Config>::AccountId {
+    <T as frame_system::Config>::AccountId::decode(&mut &[1u8; 32][..]).unwrap()
+}
+
+pub fn bob<T: frame_system::Config>() -> <T as frame_system::Config>::AccountId {
+    <T as frame_system::Config>::AccountId::decode(&mut &[2u8; 32][..]).unwrap()
+}
+
+pub fn charlie<T: frame_system::Config>() -> <T as frame_system::Config>::AccountId {
+    <T as frame_system::Config>::AccountId::decode(&mut &[3u8; 32][..]).unwrap()
+}
+
+pub fn dave<T: frame_system::Config>() -> <T as frame_system::Config>::AccountId {
+    <T as frame_system::Config>::AccountId::decode(&mut &[4u8; 32][..]).unwrap()
+}
+
+pub fn generate_account<T: frame_system::Config>(
+    seed: u32,
+) -> <T as frame_system::Config>::AccountId {
+    let mut adr = [0u8; 32];
+
+    let mut value = seed;
+    let mut id = 0;
+    while value != 0 {
+        adr[31 - id] = (value % 256) as u8;
+        value = value / 256;
+        id += 1;
+    }
+
+    <T as frame_system::Config>::AccountId::decode(&mut &adr[..]).unwrap()
+}
